requirements.txt
django
djangorestframework
requests
djangorestframework-simplejwt
django-cors-headers
python-dotenv
sqlparse
timedelta